@@ -1,35 +1,28 @@
 {
-    "name": "cakedc/users-example",
-    "description": "CakeDC/Users example",
-    "homepage": "https://github.com/CakeDC/users-example",
+    "name": "cakephp/app",
+    "description": "CakePHP skeleton app",
+    "homepage": "https://cakephp.org",
     "type": "project",
     "license": "MIT",
     "require": {
-        "php": ">=5.6",
-        "cakedc/users": "dev-develop",
-        "cakephp/cakephp": "3.7.*",
-        "cakephp/migrations": "^2.0.0",
-        "cakephp/plugin-installer": "^1.0",
-        "google/recaptcha": "@stable",
-        "league/oauth1-client": "@stable",
-        "league/oauth2-facebook": "@stable",
-        "league/oauth2-google": "@stable",
-        "league/oauth2-instagram": "@stable",
-        "league/oauth2-linkedin": "@stable",
-        "mobiledetect/mobiledetectlib": "2.*",
-        "robthree/twofactorauth": "^1.5.2"
+        "php": ">=7.2",
+        "cakephp/cakephp": "~4.1.0",
+        "cakephp/migrations": "^3.0",
+        "cakephp/plugin-installer": "^1.2",
+        "mobiledetect/mobiledetectlib": "^2.8"
     },
     "require-dev": {
-        "cakephp/bake": "^1.9.0",
-        "cakephp/cakephp-codesniffer": "^3.0",
-        "cakephp/debug_kit": "^3.17.0",
-        "josegonzalez/dotenv": "3.*",
-        "phpunit/phpunit": "^5|^6",
+        "cakephp/bake": "^2.0.3",
+        "cakephp/cakephp-codesniffer": "~4.2.0",
+        "cakephp/debug_kit": "^4.1",
+        "josegonzalez/dotenv": "^3.2",
+        "phpunit/phpunit": "~8.5.0",
         "psy/psysh": "@stable"
     },
     "suggest": {
         "markstory/asset_compress": "An asset compression plugin which provides file concatenation and a flexible filter system for preprocessing and minification.",
-        "dereuromark/cakephp-ide-helper": "After baking your code, this keeps your annotations in sync with the code evolving from there on for maximum IDE and PHPStan compatibility."
+        "dereuromark/cakephp-ide-helper": "After baking your code, this keeps your annotations in sync with the code evolving from there on for maximum IDE and PHPStan/Psalm compatibility.",
+        "phpstan/phpstan": "PHPStan focuses on finding errors in your code without actually running it. It catches whole classes of bugs even before you write tests for the code."
     },
     "autoload": {
         "psr-4": {
@@ -50,8 +43,9 @@
             "@test",
             "@cs-check"
         ],
-        "cs-check": "phpcs --colors -p --standard=vendor/cakephp/cakephp-codesniffer/CakePHP src/ tests/",
-        "cs-fix": "phpcbf --colors --standard=vendor/cakephp/cakephp-codesniffer/CakePHP src/ tests/",
+        "cs-check": "phpcs --colors -p  src/ tests/",
+        "cs-fix": "phpcbf --colors -p src/ tests/",
+        "stan": "phpstan analyse src/",
         "test": "phpunit --colors=always"
     },
     "prefer-stable": true,
